Add a message after consolidating the analysis result
On the one hand, this message is after an important step in the
build process, to giving the user a better insight into what is
going on. On the other hand, the size of the discovered graph is
useful information, e.g., when comparing with the number of actions
actually traversed when building the requested artifacts.