tool: at the earliest possible moment confirm the target
In this way the user gets an early feedback about the target that
was requested and can check if that was the target they had in
mind, especially in the case of fall back to the alphabetically
first one (according to native byte order). As a nice side effect,
we have a timestap on when the analysis started.