["", "exported-just"] make "CC" flexible
While our own tool is entirely written in C++, we link against
libraries written in C which, in the default multi-repo setup we
build from source. Therefore, the target also depends on the C
compiler, not only the C++ one. Hence also allow the user to user
a different compiler here as well, in the same way as they can
override the C++ compiler.