refactor EntityName
EntityName now clearly expresses its double identity:
 - NamedTarget
 - AnonymousTarget

The usage of std::variant<NamedTarget,AnonymousTarget> guarantees that
EntityName, internally, is not a mix of the two - like could happen
before this patch.

NamedTarget features an enum ReferenceType to express the type of the
target, namely, "normal target" or an "explicit file reference".

Thanks to this refactoring, the introduction of new targets type
should be easier, since the design is more modular.

NamedTarget