import git2: close header directory as a tree
We assume that the git2 directory in the include-file path is
owned by git2 and no other targets put files there. Document this
by closing it up as a tree. (As a side effect we pass around fewer
identifiers and reported runfiles list get easier to read.) Note
that for the tree we can't use the TREE constructor, as, besides
the source files, there is also a generated file contained in that
directory.