just-mr: more safely create directories
While the just binary itself was carefully created to work correctly
in the presence of several instances running simultaneously, the
just-mr.py script started as proof-of-concept out of the need to
conveniently generate a multi-repository configuration for just.
Nevertheless, it is now actually used, so parallel running instances
have to be taken seriously. Moreover, we have to ensure that in
case of failures, we don't leave the local build root in a state
giving a false impression on how an unpacked archive looks like.
Both can be avoided by always working on temporary directories and,
should a persistent location be needed, only renaming to the final
desitantion once everything is set up correclty.