just-mr: retry `git tag -f` if needed
just-mr uses git tags to ensure that git roots used in the repository
configuration handed out do not get garbage collected. Here, the
tag encodes the commit to be kept; hence we can safely do this
operation forcefully: we would only reset the tag to its old value.
However, this tagging still is not free of races: git seems to
first check for the existence of the tag and if not present tries
creation under the assumption that no other process does so (and
fails if this is the case). As our tagging is idempotent, we can
safely retry it to work around this git race.