Artifact reporting: repeat the failed artifacts
In targets that produce a large list of artifacts, e.g., a simple
test suite realized by "install", it is sometimes hard to find the
failed artifacts after being made aware of them by the warning.
Therefore, at info level, repeat a list of the failed artifacts
close to the end of the output.