dump-targets: don't dump source trees
We have to include explicit tree references into our target-result
map, as these targets can provide a tree definition, if the root
is not a git root or compatible mode is used. Nevertheless, from a
user's persepective, trees are just source references, like files.
Therefore, do not include them, when dumping the map of analyzed
targets. In this way, we also avoid the ambiguity in the dump between
an explicit tree reference and a defined target with the same name.