fix the path of the local tree-cas for compatible mode
in daca274041e31636f08927b0cebc56fd33c7bbdf it has been added a local
tree-cas. This cas is populated only when just runs in native
mode. However, even in compatible mode, a tree must be looked for in
the tree-cas. This patch set the path of the tree-cas to be the same
of file-cas while running in compatible mode.