just-mr.py: for the fetch subcommand acutally fetch
A checkout does not necessarily fetch an archive: if we already
have the git tree for that archive, this is enough to create a
build root. For the fetch command, however, we need to have the
actual archive as we have to copy it to the distdir. Of course, we
only create the git tree once we got hold of the archive. However,
with the introduction of CAS purging, the invariant that we have
the archive whenever we can create a build root will no longer be
true. This is acutally a feature as the git trees can be stored
more compactly if we accumulate different release archives of the
same upstream project. However, it also means that we explicitly
have to fetch the archive in the fetch subcommand. Do this.