import of boringssl: fix export structure
Both ["", "ssl"] and ["", "crypto"] are targets of "boringssl"
that are intended for use outside this repository. Both depend on
the ["CC", "library"]-target ["", "crytpo-lib"], in fact even visibly,
as ssl has a library-dependency on the crypto library. Therefore,
we have to follow the standard export pattern that all uses of this
library go through the same (closest) export target. Fix this!