just-mr: fix distdir computation
just-mr is designed to store everything that can be reported as git
tree entirely in git only. To avoid recomputation, just-mr keeps index
files mapping the description of a repository to the corresponding
git trees. For these index files to be useful, the computation of
the key has to be independent of the presence of the respective
archives in CAS. This will become even more important once garbage
collection will be added. Fix this for distdir repositories.