Fix wrong assembly of just args in just-mr
The actual error fixed in this commit was the usage of the std::vector resize
method instead of the reserve method. The resize method increases the size of
the vector by appending new default-allocated elements (empty strings in this
case), whereas the reserve method just increases the underlying capacity of the
vector without appending elements. In addition to this fix, the remaining
just-args assembly code was refactored to drop the lambda-function usage in
favor of a plain sequence of statements.