just-mr: Add git binary option
In the rare cases that we need to shell out to git, let the user
configure what binary to use. Option resolves in the same way as
the just executable, including allowing it to be set via just-mrrc.

Updates all cases of shelling out to git (fetch and commit update).

Update just-mr and just-mrrc docs accordingly.