bootstrap: Do not prune non-local-deps
Dependencies (type archive/zip) that are not required for
bootstrapping just are usually marked with the hint "drop",
in order to avoid fetching non-required archives. This
affects only non-local dependencies, as local dependencies
(from the system) will not fetch anything and therefore do
not need to be dropped.

With this commit, the "drop" hint for declared
non-local-deps will be ignored during package builds. In
this way, the user can specify dependencies that should not
be dropped (despite not strictly required to bootstrap
just). Consequently, those dependencies remain in the
generated bootstrap configuration, in case the user wants to
use this configuration to build other targets, such as
just-mr.