test: Add checks for 'special' pragma in just-mr
Checks that the new 'special' pragma works for all supported
repository types. In particular, checks that symlinks get resolved
as expected and that resolved trees add their unresolved
counterparts to the CAS for 'file' and 'archive' type repositories.

Splits install-roots test into a basic check which works with the
bootstrappable just-mr, and a new check focusing on the newly
introduced 'special' pragma.