just-mr fetch: Fix exception on fetched data hash mismatch
After successfully fetching data over the network for an archive,
if the optional checksums are not provided, the code will throw if
there is a mismatch between the hash of the fetched data stream
and the provided content hash.

This commit fixes the issue by adding an additional check which
properly handles the possible mismatch.

(cherry-picked from 9acaa7f60c88c97f58b757ffb6ce206f3f2953a2)