FileRoot: Add new absent root underlying type variant
Absent roots are characterised only by a Git tree hash, so a new
variant of the underlying stored information was added in the form
of a plain string.

In order to avoid unwanted implicit conversions when instantiating
via literal strings, we force callers of the constructors to
explicitly differentiate between plain strings and filesystem
paths. Existing tests were updated to reflect this.

Co-authored-by: Alberto Sartori <alberto.sartori@huawei.com>