Fetch export targets sequentially when creating local backup
Typically, the number of export targets is small compared to the
size of the respective export targets. Moreover, export targets are
often chained, resulting in overlapping artifacts that we want to
fetch only once. Therefore, fetch export targets sequentially, giving
room to introduce parallelism in the individual fetch steps later.

While there, also log the beginning of the artifact synchronisation
for the export targets.