remote api: support otpional parallel CAS synchronisation
Allow implementations to use a given number of threads to carry out
the synchronisation. In this way parallelism can be achieved even
in situations where batch reading degrades as objects of unknown
size have to be fetched; this situation typically occurs if a tree
object has a large number of direct children that are blobs.

If no implementation is provided, the default implementation is to
fall back to the normal (sequential) CAS synchronisation.