amend service-target-cache.md
Remove wrong (un-needed) design specifications. When we compute the
TargetCacheKey of an export target, there is no need to check if any
of (recursively) referenced trees is absent. As soon as we deal with
an "absent target" we query just serve. If it is not able to provide
the target cache value, we error out.