Test ["end-to-end/gc", "export"]: fix workspace layout
This end-to-end test, of course, depends on the tools, provided
inside the test's working directory. The test, being end to end,
constructs a workspace with multi-repository configuration with
the main repo a file repsoitory at the root of the constructed
work space; to be able to test exporting this repository has pragma
"to_git". Move the constructed work space to sub directory to not
include the tools to be tested in the work space that is gitified;
one the one hand, the tested tools should not be part of the test
workspace and on the other hand, this shaves off a large fraction
of the test's run time.