remote API: support uploading files from the file system
Allow implementations to have a specialized way of uploading a file
owned by the build process to the applicable cas, both as file, as
well as as tree object (in native mode).

If no implementation is provided, the default is to read the file
into memory and use the Upload method.